WIP: Move Repositories to Domain layer for avoid reference between Application and Domain layers
