+++ CarFactory.Employees.Application/CarFactory.Employees.Application.csproj
@@ -11,7 +11,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\CarFactory.Employees.Infrastructure\CarFactory.Employees.Infrastructure.csproj" />
+    <ProjectReference Include="..\CarFactory.Employees.Domain\CarFactory.Employees.Domain.csproj" />
     <ProjectReference Include="..\CarFactory.Employees.SharedLibrary\CarFactory.Employees.SharedLibrary.csproj" />
   </ItemGroup>
