+++ CarFactory.Employees.Application/Features/EmployeeRequests/Queries/GetEmployeeRequestsQueryHandler.cs
@@ -1,5 +1,5 @@
 ﻿using CarFactory.Employees.Application.Features.EmployeeRequests.DTOs;
-using CarFactory.Employees.Infrastructure.Repositories;
+using CarFactory.Employees.Domain.Repositories;
 using MediatR;
 
 namespace CarFactory.Employees.Application.Features.EmployeeRequests.Queries;
