+++ CarFactory.Employees.Domain/Repositories/IEmployeeRequestCandidateRepository.cs
@@ -1,0 +1,7 @@
+﻿using CarFactory.Employees.Domain.Models;
+
+namespace CarFactory.Employees.Domain.Repositories;
+
+public interface IEmployeeRequestCandidateRepository : IBaseRepository<EmployeeRequestCandidate>
+{
+}
