+++ CarFactory.Employees.Infrastructure/Repositories/BaseRepository.cs
@@ -1,18 +1,8 @@
 ﻿using CarFactory.Employees.Domain.Models;
+using CarFactory.Employees.Domain.Repositories;
 using Microsoft.EntityFrameworkCore;
 
 namespace CarFactory.Employees.Infrastructure.Repositories;
-
-public interface IBaseRepository<T> where T : BaseEntity
-{
-    Task<IEnumerable<T>> GetAllAsync(CancellationToken token);
-    Task<T?> GetByIdAsync(Guid id, CancellationToken token);
-    Task AddAsync(T entity, CancellationToken token);
-    void Update(T entity);
-    void Delete(T entity);
-    Task SaveChangesAsync(CancellationToken token);
-    Task<bool> ExistsByIdAsync(Guid id, CancellationToken token);
-}
 
 public class BaseRepository<T> : IBaseRepository<T> where T : BaseEntity
 {
