+++ CarFactory.Employees.Infrastructure/Repositories/EmployeeRequestCandidateRepository.cs
@@ -1,11 +1,7 @@
 ﻿using CarFactory.Employees.Domain.Models;
-using Microsoft.EntityFrameworkCore;
+using CarFactory.Employees.Domain.Repositories;
 
 namespace CarFactory.Employees.Infrastructure.Repositories;
-
-public interface IEmployeeRequestCandidateRepository : IBaseRepository<EmployeeRequestCandidate>
-{
-}
 
 public class EmployeeRequestCandidateRepository : BaseRepository<EmployeeRequestCandidate>, IEmployeeRequestCandidateRepository
 {
